+++ ruby-gem/spec/system/models/destroy_event_spec.rb
@@ -5,14 +5,12 @@ describe "models destroy event" do
   let!(:task2) { create :task, user: }
   let!(:user) { create :user }
 
-  it "reacts on destroy events" do
+  it "reacts on destroy events for subscribed models in an array" do
     login_as user
     visit models_destroy_event_path
     wait_for_selector ".task-row[data-task-id='#{task1.id}']"
     wait_for_selector ".task-row[data-task-id='#{task2.id}']"
-
-    # Give some time for ActionCable to connect
-    sleep 1
+    wait_for_action_cable_to_connect
 
     task1.destroy!